Bind component view to component and make non-public properties and methods available in view only
