+++ tests/Unit/views/show-name.blade.php
@@ -1,2 +1,2 @@
 <!-- Test comment <div>Commented out code</div> -->
-<span>{{ $name }}</span>
+<span>{{ $name ?? '' }}</span>
